// spring-ai-alibaba-studio/spring-ai-alibaba-studio-server/frontend/packages/spark-flow/src/demos/flow-demo/types/flow.ts
export interface ITryCatchConfig {
  /* Exception handling strategy, defaultValue: default value, failBranch: failure branch, noop: do nothing */
  strategy: 'defaultValue' | 'failBranch' | 'noop';
  /* If strategy is defaultValue, default_value needs to be defined */
  default_values?: Omit<INodeDataInputParamItem, 'value_from'>[];
}